+++ src/commands/install.rs
@@ -5,48 +5,66 @@ use crate::{
     load_manifest,
     utils::{
         copy_dir_all,
+        dryrun::get_dry_run,
         paths::{
             check_path_dir, check_path_file, d_packages, datalocalutpm, get_current_dir,
             get_ssh_dir,
         },
         specs::Extra,
-        state::{Error, ErrorKind, Result},
+        state::{Result, UtpmError},
     },
+    utpm_log,
 };
 
 use git2::{build::RepoBuilder, Cred, FetchOptions, RemoteCallbacks};
-use tracing::{debug, instrument};
-use typst_project::{heuristics::MANIFEST_FILE, manifest::Manifest};
+use tracing::instrument;
 
 use super::{link, InstallArgs};
 
+/// Entry point for the install command.
+///
+/// Cleans up temporary directories before starting the installation process.
 #[instrument]
-pub fn run(cmd: &InstallArgs) -> Result<bool> {
+pub async fn run(cmd: &InstallArgs) -> Result<bool> {
     let path = format!("{}/tmp", datalocalutpm()?);
-    if check_path_dir(&path) {
+    if check_path_dir(&path) && !get_dry_run() {
         fs::remove_dir_all(path)?;
     }
-    init(cmd, 0)?;
+    if !get_dry_run() {
+        init(cmd, 0).await?;
+    } else {
+        utpm_log!(warn, "Dry-run, can't do anything");
+    }
     Ok(true)
 }
 
-#[instrument(skip(cmd))]
-pub fn init(cmd: &InstallArgs, i: usize) -> Result<bool> {
+/// Installs dependencies recursively.
+///
+/// If a URL is provided, it clones the repository from the given URL.
+/// Otherwise, it installs dependencies from the `typst.toml` manifest in the current directory.
+pub async fn init(cmd: &InstallArgs, i: usize) -> Result<bool> {
+    utpm_log!(trace, "executing init command for install");
+
+    // Determine the source path for the installation.
     let path = if let Some(url) = &cmd.url {
+        // If a URL is provided, create a temporary directory for cloning.
         let dir = format!("{}/tmp/{}", datalocalutpm()?, i);
-        debug!("url is set to {}, creating {}", url, dir);
+        utpm_log!(debug, "url is set to {}, creating {}", url, dir);
         dir
     } else {
+        // Otherwise, use the current directory.
         let dir = get_current_dir()?;
-        debug!("url is none, current dir: {}", dir);
+        utpm_log!(debug, "url is none, current dir: {}", dir);
         dir
     };
 
+    // If a URL is provided, clone or copy the repository.
     if let Some(x) = &cmd.url {
         fs::create_dir_all(&path)?;
         let sshpath = get_ssh_dir()?;
         let ed: String = sshpath.clone() + "/id_ed25519";
         let rsa: String = sshpath + "/id_rsa";
+        // Determine the SSH key to use.
         let val: String = match env::var("UTPM_KEYPATH") {
             Ok(val) => val,
             Err(_) => {
@@ -57,6 +75,7 @@ pub fn init(cmd: &InstallArgs, i: usize) -> Result<bool> {
                 }
             }
         };
+        // Handle git and http(s) URLs.
         if x.starts_with("git") || x.starts_with("http") {
             let mut callbacks = RemoteCallbacks::new();
             callbacks.credentials(|_, username_from_url, _| {
@@ -85,23 +104,25 @@ pub fn init(cmd: &InstallArgs, i: usize) -> Result<bool> {
             builder.fetch_options(fo);
             builder.clone(&x, Path::new(&path))?;
         } else {
+            // Handle local paths.
             copy_dir_all(&x, &path)?;
         }
     };
 
-    let typstfile = path.clone() + "/" + MANIFEST_FILE;
+    // Check for a manifest file in the source directory.
+    let typstfile = path.clone() + "/typst.toml";
     if !check_path_file(&typstfile) {
         let origin = cmd.url.clone().unwrap_or("/".into());
-        println!("{}", format!("x {}", origin));
-        return Ok(false);
+        utpm_log!("{}", format!("x {}", origin));
+        return Ok::<bool, UtpmError>(false);
     }
+
+    // Load the manifest and extract UTPM-specific configurations.
     let file = load_manifest!(&path);
-    let utpm = if let Some(value) = file.tool {
-        value.get_section("utpm")?.unwrap_or(Extra::default())
-    } else {
-        Extra::default()
-    };
+    let utpm = Extra::from(file.tool);
     let namespace = utpm.namespace.unwrap_or("local".into());
+
+    // Check if the package is already installed.
     if check_path_dir(&format!(
         "{}/{}/{}/{}",
         d_packages()?,
@@ -109,43 +130,43 @@ pub fn init(cmd: &InstallArgs, i: usize) -> Result<bool> {
         &file.package.name,
         &file.package.version
     )) {
-        println!(
+        utpm_log!(
             "{}",
             format!("~ {}:{}", file.package.name, file.package.version)
         );
         return Ok(true);
     }
 
-    println!("{}", format!("Installing {}...", file.package.name));
+    // Recursively install dependencies.
+    utpm_log!("{}", format!("Installing {}...", file.package.name));
     if let Some(vec_depend) = utpm.dependencies {
         let mut y = 0;
-        vec_depend
-            .iter()
-            .map(|a| -> Result<bool> {
-                y += 1;
-                let ins = InstallArgs {
-                    force: cmd.force,
-                    url: Some(a.to_string()),
-                };
-                init(&ins, i * vec_depend.len() + y)?;
-                Ok(true)
-            })
-            .collect::<Result<Vec<bool>>>()?;
+
+        for a in &vec_depend {
+            y += 1;
+            let ins = InstallArgs {
+                force: cmd.force,
+                url: Some(a.to_string()),
+            };
+            let ini = init(&ins, i * vec_depend.len() + y);
+            let _ = Box::pin(ini).await?;
+        }
     }
+
+    // Link the installed package and clean up temporary files.
     if !cmd.url.is_none() {
         let lnk = LinkArgs {
             force: cmd.force,
             no_copy: false,
         };
-        link::run(&lnk, Some(path.clone()), false)?; //TODO: change here too
+        link::run(&lnk, Some(path.clone()), false).await?; //TODO: change here too
         fs::remove_dir_all(&path)?;
-        println!(
-            "{}",
-            format!("+ {}:{}", file.package.name, file.package.version)
-        );
+        utpm_log!(info, "+ {}:{}", file.package.name, file.package.version);
     } else {
-        println!("* Installation complete! If you want to use it as a lib, just do a `utpm link`!")
+        utpm_log!(
+            info,
+            "* Installation complete! If you want to use it as a lib, just do a `utpm link`!"
+        )
     }
-
     Ok(true)
 }
