+++ src/commands/unlink.rs
@@ -1,27 +1,31 @@
 use inquire::Confirm;
 use std::fs;
-use tracing::{info, instrument};
+use tracing::instrument;
 
 use crate::{
     format_package,
     utils::{
-        paths::{c_packages, d_packages},
-        regex_namespace, regex_package, regex_packagename,
-        state::{Error, ErrorKind, Result},
+        dryrun::get_dry_run, paths::check_path_dir, regex_namespace, regex_package,
+        regex_packagename, state::Result,
     },
+    utpm_bail, utpm_log,
 };
 
 use super::UnlinkArgs;
 
+/// Unlinks/deletes a package from the local storage.
 #[instrument(skip(cmd))]
-pub fn run(cmd: &UnlinkArgs) -> Result<bool> {
+pub async fn run(cmd: &UnlinkArgs) -> Result<bool> {
+    utpm_log!(trace, "executing unlink command");
     let packages = &cmd.package;
 
-    // RegEx
+    // Use regex to parse the package string, which can be a full package spec,
+    // a package name, or just a namespace.
     let re_all = regex_package();
     let re_name = regex_packagename();
     let re_namespace = regex_namespace();
     let path: String;
+
     if let Some(cap) = re_all.captures(packages.as_str()) {
         let (_, [namespace, package, major, minor, patch]) = cap.extract();
         path = format_package!(namespace, package, major, minor, patch);
@@ -32,22 +36,34 @@ pub fn run(cmd: &UnlinkArgs) -> Result<bool> {
         let (_, [namespace]) = cap.extract();
         path = format_package!(namespace);
     } else {
-        return Err(Error::empty(ErrorKind::PackageNotValid));
+        utpm_bail!(PackageNotValid);
     }
 
+    // Check if the package directory exists.
+    if !check_path_dir(&path) {
+        utpm_bail!(PackageNotExist)
+    }
+
+    // Confirm with the user before deleting, unless `--yes` is provided.
     if !cmd.yes {
         match Confirm::new("Are you sure to delete this? This is irreversible.")
             .with_help_message(format!("You want to delete {packages}").as_str())
             .prompt()
         {
             Ok(_) => {
-                fs::remove_dir_all(path)?;
+                utpm_log!(info, "Deleting {}", path);
+                if !get_dry_run() {
+                    fs::remove_dir_all(path)?;
+                }
                 Ok(true)
             }
             Err(_) => Ok(false),
         }
     } else {
-        fs::remove_dir_all(path)?;
+        utpm_log!(info, "Deleting {}", path);
+        if !get_dry_run() {
+            fs::remove_dir_all(path)?;
+        }
         Ok(true)
     }
 }
